+++ src/main/java/org/jhipster/blog/web/rest/EntryResource.java
@@ -50,10 +50,14 @@ public class EntryResource {
      */
     @PostMapping("/entries")
     @Timed
-    public ResponseEntity<Entry> createEntry(@Valid @RequestBody Entry entry) throws URISyntaxException {
+    public ResponseEntity<?> createEntry(@Valid @RequestBody Entry entry) throws URISyntaxException {
         log.debug("REST request to save Entry : {}", entry);
         if (entry.getId() != null) {
             throw new BadRequestAlertException("A new entry cannot already have an ID", ENTITY_NAME, "idexists");
+        }
+        if (entry.getBlog() != null &&
+            !entry.getBlog().getUser().getLogin().equals(SecurityUtils.getCurrentUserLogin().orElse(""))) {
+            return new ResponseEntity<>("Unauthorized", HttpStatus.UNAUTHORIZED);
         }
         Entry result = entryRepository.save(entry);
         return ResponseEntity.created(new URI("/api/entries/" + result.getId()))
@@ -72,10 +76,14 @@ public class EntryResource {
      */
     @PutMapping("/entries")
     @Timed
-    public ResponseEntity<Entry> updateEntry(@Valid @RequestBody Entry entry) throws URISyntaxException {
+    public ResponseEntity<?> updateEntry(@Valid @RequestBody Entry entry) throws URISyntaxException {
         log.debug("REST request to update Entry : {}", entry);
         if (entry.getId() == null) {
             throw new BadRequestAlertException("Invalid id", ENTITY_NAME, "idnull");
+        }
+        if (entry.getBlog() != null &&
+            !entry.getBlog().getUser().getLogin().equals(SecurityUtils.getCurrentUserLogin().orElse(""))) {
+            return new ResponseEntity<>("Unauthorized", HttpStatus.UNAUTHORIZED);
         }
         Entry result = entryRepository.save(entry);
         return ResponseEntity.ok()
@@ -109,9 +117,13 @@ public class EntryResource {
      */
     @GetMapping("/entries/{id}")
     @Timed
-    public ResponseEntity<Entry> getEntry(@PathVariable Long id) {
+    public ResponseEntity<?> getEntry(@PathVariable Long id) {
         log.debug("REST request to get Entry : {}", id);
         Optional<Entry> entry = entryRepository.findOneWithEagerRelationships(id);
+        if (entry.isPresent() && entry.get().getBlog() != null &&
+            !entry.get().getBlog().getUser().getLogin().equals(SecurityUtils.getCurrentUserLogin().orElse(""))) {
+            return new ResponseEntity<>("Unauthorized", HttpStatus.UNAUTHORIZED);
+        }
         return ResponseUtil.wrapOrNotFound(entry);
     }
 
@@ -123,9 +135,13 @@ public class EntryResource {
      */
     @DeleteMapping("/entries/{id}")
     @Timed
-    public ResponseEntity<Void> deleteEntry(@PathVariable Long id) {
+    public ResponseEntity<?> deleteEntry(@PathVariable Long id) {
         log.debug("REST request to delete Entry : {}", id);
-
+        Optional<Entry> entry = entryRepository.findOneWithEagerRelationships(id);
+        if (entry.isPresent() && entry.get().getBlog() != null &&
+            !entry.get().getBlog().getUser().getLogin().equals(SecurityUtils.getCurrentUserLogin().orElse(""))) {
+            return new ResponseEntity<>("Unauthorized", HttpStatus.UNAUTHORIZED);
+        }
         entryRepository.deleteById(id);
         return ResponseEntity.ok().headers(HeaderUtil.createEntityDeletionAlert(ENTITY_NAME, id.toString())).build();
     }
